+++ spec/property.spec.ts
@@ -1,14 +1,15 @@
+import 'reflect-metadata';
 import { Alias, Property, Serializable, Deserializable, TypeArtisan, DefaultValue, UseDefaultValue, UseImplicitConversion } from './../src';
 
 class X
 {
     @Property() public a?: string;
-    @Property() @Alias('e') public b?: string;
-    @Property() @Serializable() public c?: string;
-    @Property() @Deserializable() public d?: string;
-    @Property() @DefaultValue('e') public e?: string;
-    @Property() @UseDefaultValue() public f?: string;
-    @Property() @UseImplicitConversion() public g?: string;
+    @Property('String') @Alias('e') public b?: string;
+    @Property(() => String) @Serializable() public c?: string;
+    @Property() @Deserializable() public d?: number;
+    @Property('Number') @DefaultValue(1) public e?: number;
+    @Property(() => Number) @UseDefaultValue() public f?: number;
+    @Property() @UseImplicitConversion() public g?: boolean;
 }
 
 describe('Property decorator', function () 
@@ -40,6 +41,9 @@ describe('Property decorator', function ()
         expect(aPropertyMetadata?.defaultValue).not.toBeDefined();
         expect(aPropertyMetadata?.useDefaultValue).not.toBeDefined();
         expect(aPropertyMetadata?.useImplicitConversion).not.toBeDefined();
+        expect(aPropertyMetadata?.typeResolver).toBeUndefined();
+        expect(aPropertyMetadata?.reflectTypeResolver).toBeDefined();
+        expect(aPropertyMetadata?.reflectTypeResolver()).toBe(String);
 
         expect(bPropertyMetadata).toBeDefined();
         expect(bPropertyMetadata?.name).toBe('b');
@@ -49,6 +53,9 @@ describe('Property decorator', function ()
         expect(bPropertyMetadata?.defaultValue).not.toBeDefined();
         expect(bPropertyMetadata?.useDefaultValue).not.toBeDefined();
         expect(bPropertyMetadata?.useImplicitConversion).not.toBeDefined();
+        expect(bPropertyMetadata?.typeResolver).toBeDefined();
+        expect(bPropertyMetadata?.reflectTypeResolver).toBeDefined();
+        expect(bPropertyMetadata?.reflectTypeResolver()).toBe(String);
 
         expect(cPropertyMetadata).toBeDefined();
         expect(cPropertyMetadata?.name).toBe('c');
@@ -58,6 +65,9 @@ describe('Property decorator', function ()
         expect(cPropertyMetadata?.defaultValue).not.toBeDefined();
         expect(cPropertyMetadata?.useDefaultValue).not.toBeDefined();
         expect(cPropertyMetadata?.useImplicitConversion).not.toBeDefined();
+        expect(cPropertyMetadata?.typeResolver).toBeDefined();
+        expect(cPropertyMetadata?.reflectTypeResolver).toBeDefined();
+        expect(cPropertyMetadata?.reflectTypeResolver()).toBe(String);
 
         expect(dPropertyMetadata).toBeDefined();
         expect(dPropertyMetadata?.name).toBe('d');
@@ -67,15 +77,21 @@ describe('Property decorator', function ()
         expect(dPropertyMetadata?.defaultValue).not.toBeDefined();
         expect(dPropertyMetadata?.useDefaultValue).not.toBeDefined();
         expect(dPropertyMetadata?.useImplicitConversion).not.toBeDefined();
+        expect(dPropertyMetadata?.typeResolver).toBeUndefined();
+        expect(dPropertyMetadata?.reflectTypeResolver).toBeDefined();
+        expect(dPropertyMetadata?.reflectTypeResolver()).toBe(Number);
 
         expect(ePropertyMetadata).toBeDefined();
         expect(ePropertyMetadata?.name).toBe('e');
         expect(ePropertyMetadata?.alias).not.toBeDefined();
         expect(ePropertyMetadata?.serializable).not.toBeDefined();
         expect(ePropertyMetadata?.deserializable).not.toBeDefined();
-        expect(ePropertyMetadata?.defaultValue).toBe('e');
+        expect(ePropertyMetadata?.defaultValue).toBe(1);
         expect(ePropertyMetadata?.useDefaultValue).not.toBeDefined();
         expect(ePropertyMetadata?.useImplicitConversion).not.toBeDefined();
+        expect(ePropertyMetadata?.typeResolver).toBeDefined();
+        expect(ePropertyMetadata?.reflectTypeResolver).toBeDefined();
+        expect(ePropertyMetadata?.reflectTypeResolver()).toBe(Number);
 
         expect(fPropertyMetadata).toBeDefined();
         expect(fPropertyMetadata?.name).toBe('f');
@@ -85,6 +101,9 @@ describe('Property decorator', function ()
         expect(fPropertyMetadata?.defaultValue).not.toBeDefined();
         expect(fPropertyMetadata?.useDefaultValue).toBeTrue();
         expect(fPropertyMetadata?.useImplicitConversion).not.toBeDefined();
+        expect(fPropertyMetadata?.typeResolver).toBeDefined();
+        expect(fPropertyMetadata?.reflectTypeResolver).toBeDefined();
+        expect(fPropertyMetadata?.reflectTypeResolver()).toBe(Number);
 
         expect(gPropertyMetadata).toBeDefined();
         expect(gPropertyMetadata?.name).toBe('g');
@@ -94,5 +113,8 @@ describe('Property decorator', function ()
         expect(gPropertyMetadata?.defaultValue).not.toBeDefined();
         expect(gPropertyMetadata?.useDefaultValue).not.toBeDefined();
         expect(gPropertyMetadata?.useImplicitConversion).toBeTrue();
+        expect(gPropertyMetadata?.typeResolver).toBeUndefined();
+        expect(gPropertyMetadata?.reflectTypeResolver).toBeDefined();
+        expect(gPropertyMetadata?.reflectTypeResolver()).toBe(Boolean);
     });
 });