+++ src/property.metadata.ts
@@ -1,4 +1,5 @@
 import { Fn } from './utils';
+import { TypeCtor } from './type.ctor';
 import { TypeResolver } from './type.resolver';
 import { TypeSerializer } from './type.serializer';
 import { PropertyOptions } from './property.options';
@@ -18,11 +19,27 @@ export class PropertyMetadata
     public name: string;
 
     /**
+     * Type resolver defined using reflect metadata.
+     * 
+     * Used as a fallback when type resolver is not defined.
+     * 
+     * @type {TypeResolver}
+     */
+    public reflectTypeResolver: TypeResolver;
+
+    /**
      * Type resolver to get a property type.
      * 
      * @type {TypeResolver}
      */
-    public typeResolver: TypeResolver;
+    public typeResolver?: TypeResolver;
+
+    /**
+     * Type alias.
+     * 
+     * @type {string}
+     */
+    public typeAlias?: string;
 
     /**
      * Custom property type serializer.
@@ -57,6 +74,13 @@ export class PropertyMetadata
     public deserializable?: boolean;
 
     /**
+     * Multiple property?
+     * 
+     * @type {boolean}
+     */
+    public multiple?: boolean;
+
+    /**
      * Default value for undefined ones.
      * 
      * Assigned only when use default value option is true.
@@ -83,12 +107,16 @@ export class PropertyMetadata
     /**
      * Constructor.
      * 
+     * @param {TypeCtor} declaringTypeCtor Declaring type constructor.
      * @param {string} name Property name.
      */
-    public constructor(name: string)
+    public constructor(declaringTypeCtor: TypeCtor, name: string)
     {
-        this.name         = name;
-        this.typeResolver = () => null;
+        const propertyTypeCtor = Fn.extractReflectMetadata('design:type', declaringTypeCtor.prototype, name);
+
+        this.name                = name;
+        this.reflectTypeResolver = () => propertyTypeCtor;
+        this.multiple            = propertyTypeCtor === Array;
 
         return;
     }
@@ -117,6 +145,11 @@ export class PropertyMetadata
             this.typeResolver = propertyOptions.typeResolver;
         }
 
+        if (!Fn.isUndefined(propertyOptions.typeAlias)) 
+        {
+            this.typeAlias = propertyOptions.typeAlias;
+        }
+
         if (!Fn.isUndefined(propertyOptions.typeSerializer)) 
         {
             this.typeSerializer = propertyOptions.typeSerializer;
@@ -135,6 +168,11 @@ export class PropertyMetadata
         if (!Fn.isUndefined(propertyOptions.deserializable))
         {
             this.deserializable = propertyOptions.deserializable;
+        }
+
+        if (!Fn.isUndefined(propertyOptions.multiple))
+        {
+            this.multiple = propertyOptions.multiple;
         }
 
         if (!Fn.isUndefined(propertyOptions.defaultValue))