+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2020-12-23
+
+### Added
+
+- Multiple property option to indicate that certain property is an array when using without reflect metadata.
+
+### Fixed
+
+- Type extraction using reflect metadata.
+
 ## [1.1.0] - 2020-12-22
 
 ### Added